+++ iti-react/index.ts
@@ -1,7 +1,6 @@
 ﻿export * from '@interface-technologies/iti-react-core'
 
 export * from './Validation'
-export * from './DataUpdater'
 export * from './Inputs'
 export * from './Components'
 export * from './Routing'