+++ test-website/src/Pages/Product/List.tsx
@@ -1,26 +1,50 @@
 ﻿import React from 'react'
-import { useState } from 'react'
 import moment from 'moment-timezone'
-import { ProductDto } from 'Models'
-import { PageProps } from 'Components/Routing/RouteProps'
-import {
-    AutoRefreshUpdater,
-    DataUpdater,
-    Pager,
-    getTdLink,
-    CancellablePromise,
-    resetPageIfFiltersChanged,
-    preventNonExistentPage,
-    useParameterizedQuery,
-    getTotalPages,
-    usePaginationHelpers,
-    useParameterizedAutoRefreshQuery
-} from '@interface-technologies/iti-react'
-import { api } from 'Api'
+import { useState } from 'react'
+import { PageProps } from 'Components/Routing'
 import { NavbarLink } from 'Components'
+import { RadioInput, RadioOption, useParameterizedQuery, useParameterizedAutoRefreshQuery, usePaginationHelpers, getTdLink, Pager } from '@interface-technologies/iti-react'
+import { ProductDto } from 'Models'
 import { QueryControlsWrapper } from 'Components/QueryControlsWrapper'
-import { isConnectionError } from '_Redux'
+import { api } from 'Api'
 
+export function Page(props: PageProps) {
+    const { onError, ready } = props
+
+    function onReady() {
+        props.onReady({
+            title: 'Products',
+            activeNavbarLink: NavbarLink.Products
+        })
+    }
+
+    const [hook, setHook] = useState<HookName>('useParameterizedQuery')
+
+    const options: RadioOption[] = hookNames.map(h => ({ value: h, label: h }))
+
+    return (
+        <div hidden={!ready}>
+            <div className="mb-4">
+                <RadioInput
+                    name="hook"
+                    value={hook}
+                    onChange={value => setHook(value as HookName)}
+                    options={options}
+                    validators={[]}
+                    showValidation={false}
+                />
+            </div>
+            <ListCore
+                hook={hook}
+                onReady={onReady}
+                onError={onError}
+                // force component to remount when hook changes since we are disobeying the rules of hooks
+                // in the component
+                key={hook}
+            />
+        </div>
+    )
+}
 interface QueryParams {
     name: string
     page: number
@@ -125,7 +149,7 @@ export function ListCore(props: ListCoreProps) {
     // We're breaking the rules of hooks here but it's OK because we force the component
     // to remount when the hook changes
     if (hook === 'useParameterizedQuery') {
-        ;({ doQuery, doQueryAsync } = useParameterizedQuery<QueryParams, QueryResult>({
+        ; ({ doQuery, doQueryAsync } = useParameterizedQuery<QueryParams, QueryResult>({
             queryParams,
             query: qp =>
                 api.product
@@ -143,56 +167,56 @@ export function ListCore(props: ListCoreProps) {
         }))
     } else if (hook === 'useParameterizedAutoRefreshQuery') {
         let startAutoRefresh: () => void
-        ;({ doQuery, doQueryAsync, startAutoRefresh } = useParameterizedAutoRefreshQuery<
-            QueryParams,
-            QueryResult
-        >({
-            queryParams,
-            query: qp =>
-                api.product
-                    .list({
-                        name: qp.name,
-                        page: qp.page,
-                        pageSize: qp.pageSize
-                    })
-                    .then(list => ({ ...list, pageSize: qp.pageSize })),
-            shouldQueryImmediately: (prev, curr) =>
-                prev.page !== curr.page || prev.pageSize !== curr.pageSize,
-            onLoadingChange: setLoading,
-            onResultReceived,
-            refreshInterval: moment.duration(5, 'seconds'),
-            onRefreshingChange: setRefreshing,
-            onConnectionError: () => setHasConnectionError(true),
-            onOtherError: onError
-        }))
-        ;(window as any).startAutoRefresh = startAutoRefresh
+            ; ({ doQuery, doQueryAsync, startAutoRefresh } = useParameterizedAutoRefreshQuery<
+                QueryParams,
+                QueryResult
+            >({
+                queryParams,
+                query: qp =>
+                    api.product
+                        .list({
+                            name: qp.name,
+                            page: qp.page,
+                            pageSize: qp.pageSize
+                        })
+                        .then(list => ({ ...list, pageSize: qp.pageSize })),
+                shouldQueryImmediately: (prev, curr) =>
+                    prev.page !== curr.page || prev.pageSize !== curr.pageSize,
+                onLoadingChange: setLoading,
+                onResultReceived,
+                refreshInterval: moment.duration(5, 'seconds'),
+                onRefreshingChange: setRefreshing,
+                onConnectionError: () => setHasConnectionError(true),
+                onOtherError: onError
+            }))
+            ; (window as any).startAutoRefresh = startAutoRefresh
     } else {
         throw new Error(`Unexpected hook: ${hook}.`)
     }
 
     // Call from the browser console to test that these methods work
-    ;(window as any).doQuery = doQuery
-    ;(window as any).doQueryAsync = doQueryAsync
+    ; (window as any).doQuery = doQuery
+        ; (window as any).doQueryAsync = doQueryAsync
 
     const totalPages = usePaginationHelpers({
-            queryParams,
-            items: products,
-            totalCount: totalFilteredCount,
-            pageSizeWhenItemsRetrieved: pageSizeWhenProductsRetrieved,
-            onPageChange: page => setQueryParams(qp => ({ ...qp, page }))
-        })
+        queryParams,
+        items: products,
+        totalCount: totalFilteredCount,
+        pageSizeWhenItemsRetrieved: pageSizeWhenProductsRetrieved,
+        onPageChange: page => setQueryParams(qp => ({ ...qp, page }))
+    })
 
         // To test that preventNonExistentPage is working correctly, edit ProductController
         // to only return 10 products. Then call this function from the browser's console,
         // go to page 2, and verify that you are automatically put back on page 1.
-    ;(window as any).setPageTo2 = () => setTotalFilteredCount(queryParams.pageSize + 1)
+        ; (window as any).setPageTo2 = () => setTotalFilteredCount(queryParams.pageSize + 1)
 
-    // To test configurable page size. Perhaps should bring ConfigurablePager into iti-react?
-    ;(window as any).setPageSize = (pageSize: number) =>
-        setQueryParams(qp => ({
-            ...qp,
-            pageSize: pageSize ? pageSize : defaultQueryParams.pageSize
-        }))
+        // To test configurable page size. Perhaps should bring ConfigurablePager into iti-react?
+        ; (window as any).setPageSize = (pageSize: number) =>
+            setQueryParams(qp => ({
+                ...qp,
+                pageSize: pageSize ? pageSize : defaultQueryParams.pageSize
+            }))
 
     const loadingClasses = ['text-primary', 'd-inline-block', 'mr-3']
     if (!loading) loadingClasses.push('invisible')